Catch 'Em All
Project Title: Catch 'Em All - Pokemon Guessing Game

Description:
Catch 'Em All is an interactive web-based game built using HTML, CSS, and JavaScript in the frontend. It utilizes MySQL as the backend database and JDBC for connectivity. The project aims to provide users with an engaging gaming experience where they can create, delete, and update their account information.

In this game, users are challenged to think of a Generation 1 Pokemon and the game utilizes a series of questions to accurately guess the chosen Pokemon with 100% accuracy. The interactive nature of the game keeps users entertained and engaged as they try to stump the system and test its ability to correctly identify their chosen Pokemon.

Key Features:

HTML, CSS, and JavaScript-based frontend for an immersive user interface.
MySQL database for efficient data storage and retrieval.
JDBC for seamless connectivity between the frontend and backend.
User account management with the ability to create, delete, and update account information.
Interactive gameplay that asks users a series of questions to guess the chosen Pokemon accurately.
With Catch 'Em All, users can enjoy a fun and challenging experience while testing their knowledge of Generation 1 Pokemon. Whether you're a Pokemon enthusiast or simply looking for an entertaining game, Catch 'Em All provides a delightful gaming experience.

Note: This project is for educational purposes only and is not affiliated with or endorsed by The Pokemon Company.